community_db_hw2/homepage_ui.py
import mysql.connector # Import MySQL connector

class UserHomepageWindow:
    def __init__(self, db_connection, username, user_avatar_path=None):
        self.db_connection = db_connection
        self.username = username
        self.user_avatar_path = user_avatar_path # Path to user's avatar image
        # self.window = ... # Initialize your UI window
        # self.window = tk.Toplevel() # Example for Tkinter
        # self.window.title(f"Homepage - {username}")
        print(f"Initializing User Homepage for {username}")
        self._setup_ui()
        self._load_sections()

    def _setup_ui(self):
        # Top Navigation Bar
        # self.nav_bar = tk.Frame(self.window) # Example for Tkinter
        # self.nav_bar.pack(fill="x")
        print("  Creating top navigation bar")

        # Welcome message (left)
        # self.welcome_label = tk.Label(self.nav_bar, text=f"Welcome, {self.username}!")
        # self.welcome_label.pack(side="left", padx=10, pady=5)
        print(f"    Welcome label: Welcome, {self.username}!")

        # User menu icon (right) - e.g., a button with user avatar
        # For simplicity, using a text button first. Avatar image requires PIL/Pillow for Tkinter.
        # self.user_menu_button = tk.Menubutton(self.nav_bar, text="User Menu", relief="raised") # Example
        # self.user_menu_button.pack(side="right", padx=10, pady=5)
        # self.user_menu = tk.Menu(self.user_menu_button, tearoff=0)
        # self.user_menu_button["menu"] = self.user_menu
        # if self.user_avatar_path:
        #     # Add logic to display avatar image (e.g., on the button or in the menu)
        #     # self.user_menu.add_command(label=f"Avatar: {self.user_avatar_path.split('/')[-1]}") # Placeholder
        #     pass
        # self.user_menu.add_command(label=f"Username: {self.username}")
        # self.user_menu.add_command(label="Personal Center", command=self.open_personal_center) # Placeholder command
        # self.user_menu.add_separator()
        # self.user_menu.add_command(label="Logout", command=self.handle_logout)
        print("    User menu icon (displays user avatar or text 'User Menu')")
        print("    User dropdown menu (configured with options)")

        # Main Content Area
        # self.main_content_area = tk.Frame(self.window) # Frame or container for sections
        # self.main_content_area.pack(fill="both", expand=True, padx=10, pady=10)
        print("  Creating main content area for sections")

        # Section List (placeholder, will be populated by _load_sections)
        # self.section_list_container = tk.Frame(self.main_content_area) # Scrollable area for section cards
        # self.section_list_container.pack(fill="both", expand=True)
        # Add a Canvas and Scrollbar for scrollable content if many sections
        print("    Section list container (card layout)")


    def toggle_user_menu(self):
        # For Tkinter Menubutton, the menu appears on click automatically.
        # If implementing a custom dropdown, this logic would show/hide it.
        print("Toggling user dropdown menu (handled by UI library for Menubutton)")
        pass
    
    def open_personal_center(self):
        print("Personal Center clicked - TBD")
        # Placeholder for opening personal center UI
        pass


    def handle_logout(self):
        print(f"User {self.username} logging out.")
        # self.window.destroy() # Close current homepage window - UI specific
        from login_ui import LoginWindow # Import here
        login_window = LoginWindow(self.db_connection)
        login_window.show()
        pass

    def _load_sections(self):
        print("Loading sections from database...")
        try:
            cursor = self.db_connection.cursor(dictionary=True) # dictionary=True for named columns
            query = "SELECT section_id, section_name, section_description FROM section"
            cursor.execute(query)
            sections = cursor.fetchall()
            cursor.close()

            if sections:
                for section_data in sections:
                    self._create_section_card(section_data)
            else:
                print("No sections found.")
                # Display a message in UI if no sections (e.g., tk.Label(self.section_list_container, text="No sections available.").pack())
        except mysql.connector.Error as err:
            print(f"Database error while loading sections: {err}")
            # Display error in UI
        except Exception as e:
            print(f"An unexpected error occurred while loading sections: {e}")
            # Display error in UI
        pass

    def _create_section_card(self, section_data):
        section_id = section_data['section_id']
        section_name = section_data['section_name']
        section_description = section_data['section_description']

        # Create a card UI element for the section
        # card_frame = tk.Frame(self.section_list_container, borderwidth=2, relief="groove", padx=5, pady=5) # Example
        # card_frame.pack(pady=5, fill="x")

        # name_label = tk.Label(card_frame, text=section_name, font=("Arial", 16, "bold"))
        # name_label.pack(anchor="w")
        # desc_label = tk.Label(card_frame, text=section_description[:100] + "...", wraplength=300, justify="left") # Truncate and wrap
        # desc_label.pack(anchor="w")

        # card_frame.bind("<Enter>", lambda e, cf=card_frame: self.on_card_hover(cf))
        # card_frame.bind("<Leave>", lambda e, cf=card_frame: self.on_card_leave(cf))
        # card_frame.bind("<Button-1>", lambda event, sid=section_id: self.open_section_detail(sid))
        # name_label.bind("<Button-1>", lambda event, sid=section_id: self.open_section_detail(sid)) # Make labels clickable too
        # desc_label.bind("<Button-1>", lambda event, sid=section_id: self.open_section_detail(sid))
        
        print(f"  Creating section card: {section_name}")
        print(f"    Description: {section_description[:30]}...") # Show truncated desc
        pass

    def on_card_hover(self, card_widget):
        # Apply hover effect (e.g., change background, shadow, scale)
        # card_widget.config(bg="lightgrey") # Example for Tkinter
        print(f"Hovering over section card.") 
        pass

    def on_card_leave(self, card_widget):
        # Remove hover effect
        # card_widget.config(bg="SystemButtonFace") # Example for Tkinter (default color)
        print(f"Leaving section card.") 
        pass

    def open_section_detail(self, section_id):
        print(f"Opening details for section_id: {section_id}")
        # Navigate to section detail page, passing section_id
        # section_detail_window = SectionDetailWindow(self.db_connection, section_id, self.username)
        # section_detail_window.show()
        # For now, just print
        pass

    def show(self):
        # self.window.mainloop() # Or similar method to show/run the UI
        print("Displaying User Homepage")

if __name__ == '__main__':
    # import mysql.connector
    # try:
    #     db_conn = mysql.connector.connect(
    #         host="localhost",
    #         user="your_db_user",
    #         password="your_db_password",
    #         database="community_database"
    #     )
    #     print("Database connected successfully for Homepage direct run.")
    #     # root = tk.Tk() # Main window for Tkinter
    #     # homepage_app = UserHomepageWindow(db_conn, "TestUser", None, root) # Pass root if needed
    #     homepage_app = UserHomepageWindow(db_conn, "TestUser", None)
    #     homepage_app.show()
    #     # root.mainloop()
    # except mysql.connector.Error as err:
    #     print(f"Failed to connect to database: {err}")
    # finally:
    #     if 'db_conn' in locals() and db_conn.is_connected():
    #         db_conn.close()
    #         print("Database connection closed.")
    print("User Homepage UI conceptual structure.")
