community_db_hw2/requirements.txt
Flask
mysql-connector-python
Pillow
Werkzeug